+++ xyui/avalonia/src/XYUI.Avalonia/Controls/XYUI1/_Shared/Styles/XyuiComponentStyles.XYUI4.cs
@@ -3,6 +3,7 @@ using Avalonia.Controls;
 using Avalonia.Markup.Xaml.MarkupExtensions;
 using Avalonia.Media;
 using Avalonia.Styling;
+using XYUI.Avalonia.Typography;
 
 namespace XYUI.Avalonia.Controls;